move compont checkout to container, add spinner
