+++ src/containers/checkout/checkout.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import OrderForm from './order-form';
-import Cart from '../cart';
+import Cart from '../../pages/cart';
 
 export const Checkout = () => (
   <div>
