+++ src/containers/checkout/delivery/delivery-type/mail-services/self-pickup/self-pickup-bottom.js
